+++ changelog.md
@@ -1,6 +1,12 @@
 # ACEmulator Change Log
 
 ### 2017-04-09
+[Ripley]
+* Added DelayedTeleport and converted MarketplaceRecall and LifestoneRecall to use it.
+* Using a lifestone recall command now uses mana as it did in retail.
+* Added animation to using a lifestone.
+
+### 2017-04-09
 [fantoms]
 * Added visual effect and sound broadcasting in action queue processing.
 * Added visual effect broadcasting to `Landblocks`.